+++ Reward/src/main/java/org/example/exceptions/UnacceptableAmountException.java
@@ -1,7 +1,9 @@
 package org.example.exceptions;
 
+import java.math.BigDecimal;
+
 public class UnacceptableAmountException extends RuntimeException {
-    public UnacceptableAmountException(Double amount) {
+    public UnacceptableAmountException(BigDecimal amount) {
         super(String.format("Unacceptable amount = %s", amount));
     }
 }
